workers/export-worker/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import uvicorn
import os
from dotenv import load_dotenv
import json
import tempfile

load_dotenv()

app = FastAPI(title="Export Worker", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class ExportRequest(BaseModel):
    content: Dict[str, Any]
    format: str  # docx, pdf, md, html

class ExportResponse(BaseModel):
    file_url: str
    file_size: int
    format: str

@app.get("/health")
async def health_check():
    return {"status": "healthy", "service": "export-worker"}

@app.post("/export", response_model=ExportResponse)
async def export_document(request: ExportRequest):
    try:
        file_url, file_size = export_to_format(request.content, request.format)
        
        return ExportResponse(
            file_url=file_url,
            file_size=file_size,
            format=request.format
        )
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

def export_to_format(content: Dict[str, Any], format_type: str) -> tuple[str, int]:
    """Export content to specified format"""
    if format_type == "docx":
        return export_to_docx(content)
    elif format_type == "pdf":
        return export_to_pdf(content)
    elif format_type == "md":
        return export_to_markdown(content)
    elif format_type == "html":
        return export_to_html(content)
    else:
        raise ValueError(f"Unsupported format: {format_type}")

def export_to_docx(content: Dict[str, Any]) -> tuple[str, int]:
    """Export to DOCX format"""
    # Placeholder implementation
    with tempfile.NamedTemporaryFile(suffix=".docx", delete=False) as tmp_file:
        # TODO: Implement actual DOCX generation
        tmp_file.write(b"DOCX content placeholder")
        file_size = tmp_file.tell()
    
    return f"file://{tmp_file.name}", file_size

def export_to_pdf(content: Dict[str, Any]) -> tuple[str, int]:
    """Export to PDF format"""
    # Placeholder implementation
    with tempfile.NamedTemporaryFile(suffix=".pdf", delete=False) as tmp_file:
        # TODO: Implement actual PDF generation
        tmp_file.write(b"PDF content placeholder")
        file_size = tmp_file.tell()
    
    return f"file://{tmp_file.name}", file_size

def export_to_markdown(content: Dict[str, Any]) -> tuple[str, int]:
    """Export to Markdown format"""
    # Placeholder implementation
    with tempfile.NamedTemporaryFile(suffix=".md", delete=False) as tmp_file:
        # TODO: Implement actual Markdown generation
        markdown_content = "# Document\n\nContent placeholder"
        tmp_file.write(markdown_content.encode())
        file_size = tmp_file.tell()
    
    return f"file://{tmp_file.name}", file_size

def export_to_html(content: Dict[str, Any]) -> tuple[str, int]:
    """Export to HTML format"""
    # Placeholder implementation
    with tempfile.NamedTemporaryFile(suffix=".html", delete=False) as tmp_file:
        # TODO: Implement actual HTML generation
        html_content = "<html><body><h1>Document</h1><p>Content placeholder</p></body></html>"
        tmp_file.write(html_content.encode())
        file_size = tmp_file.tell()
    
    return f"file://{tmp_file.name}", file_size

if __name__ == "__main__":
    port = int(os.getenv("PORT", 8007))
    uvicorn.run(app, host="0.0.0.0", port=port)
