workers/export-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
python-docx==1.1.0
reportlab==4.0.7
markdown==3.5.1
jinja2==3.1.2
